Adds Resources to README.md
I used the library for my own payment client/server implementation. Well done and thank you for it! I have written a short article on how to use it in React with an Express server. It keeps the boilerplate to a minimum, shows the straight forward implementation and offers the whole project as open source. Perhaps it is good for people using your library to find and article about it how to use it. Thanks again for your efforts!